Add Token support
This adds a user/#name/token endpoint that generates a token for access.
The rest of the endpoints validate the token if provided in the request
header, DRP-AUTH-TOKEN.  Individual endpoints will use the
encoded structure to restrict actions based upon the internal values.
The token also has an expiration value as well.

+++ frontend/isos.go
@@ -80,6 +80,9 @@ func (f *Frontend) InitIsoApi() {
 	//       404: ErrorResponse
 	f.ApiGroup.GET("/isos",
 		func(c *gin.Context) {
+			if !assureAuth(c, f.Logger, "isos", "list", "") {
+				return
+			}
 			ents, err := ioutil.ReadDir(path.Join(f.FileRoot, "isos"))
 			if err != nil {
 				c.JSON(http.StatusNotFound,
@@ -112,6 +115,9 @@ func (f *Frontend) InitIsoApi() {
 	//       404: ErrorResponse
 	f.ApiGroup.GET("/isos/:name",
 		func(c *gin.Context) {
+			if !assureAuth(c, f.Logger, "isos", "get", c.Param(`name`)) {
+				return
+			}
 			isoName := path.Join(f.FileRoot, `isos`, path.Base(c.Param(`name`)))
 			c.File(isoName)
 		})
@@ -138,6 +144,9 @@ func (f *Frontend) InitIsoApi() {
 	//       507: ErrorResponse
 	f.ApiGroup.POST("/isos/:name",
 		func(c *gin.Context) {
+			if !assureAuth(c, f.Logger, "isos", "post", c.Param(`name`)) {
+				return
+			}
 			uploadIso(c, f.FileRoot, c.Param(`name`), f.dt)
 		})
 	// swagger:route DELETE /isos/{path} Isos deleteIso
@@ -154,6 +163,9 @@ func (f *Frontend) InitIsoApi() {
 	f.ApiGroup.DELETE("/isos/:name",
 		func(c *gin.Context) {
 			name := c.Param(`name`)
+			if !assureAuth(c, f.Logger, "isos", "delete", name) {
+				return
+			}
 			isoName := path.Join(f.FileRoot, `isos`, path.Base(name))
 			if err := os.Remove(isoName); err != nil {
 				c.JSON(http.StatusNotFound,
